Add files via upload
In this notebook, a CNN model is created to learn to classify the direction of the car from a still image. The still images used for training are generated while the user (in this case me) plays the game themself. The labels are action motion the car can take. For simplicity the only actions recorded are acceleration, breaking, turn left, turn right. Motions involving multiple actions such as accelerating while turning were ignored. The action arrays were converted to integers for the model to use as labels. 

The trained model was tested on augmented images, which were also recorded at the time the user was playing the game. Another instance of the model was created and placed in the main loop of the cell which runs the game. At each timestep, the environment is passed to the model as a tensor image and the model returns a label which gets mapped to an action which moves the game to the next timestep. The process gets repeated as long as the game runs for.